AcademiaJoeDayz/requirements.txt
Django==1.6.1
Pillow==2.3.0
django-tinymce==1.5.2
mysql-python==1.2.4